docker/test-project/repo-concierge/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repo-concierge"
version = "0.1.0"
description = "A Python CLI tool that scans codebases for risky patterns, generates security audit reports, and returns meaningful exit codes for CI integration."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
repo-concierge = "repo_concierge.cli:main"

[tool.setuptools.packages.find]
include = ["repo_concierge*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
